[refactor] JwtTokenProvider 중복 제거 (#337)
