+++ backend/src/main/java/dev/tripdraw/auth/application/AuthExtractor.java
@@ -4,8 +4,8 @@ import static dev.tripdraw.auth.exception.AuthExceptionType.INVALID_AUTH_HEADER;
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static org.springframework.util.StringUtils.hasText;
 
-import dev.tripdraw.auth.dto.LoginUser;
 import dev.tripdraw.auth.exception.AuthException;
+import dev.tripdraw.common.auth.LoginUser;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
