+++ backend/src/main/java/dev/tripdraw/auth/dto/OauthInfo.java
@@ -1,6 +1,6 @@
 package dev.tripdraw.auth.dto;
 
-import dev.tripdraw.auth.domain.OauthType;
+import dev.tripdraw.common.auth.OauthType;
 
 public record OauthInfo(String oauthId, OauthType oauthType) {
 }
