+++ backend/src/main/java/dev/tripdraw/auth/dto/OauthRequest.java
@@ -1,6 +1,6 @@
 package dev.tripdraw.auth.dto;
 
-import dev.tripdraw.auth.domain.OauthType;
+import dev.tripdraw.common.auth.OauthType;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 public record OauthRequest(
