+++ backend/src/main/java/dev/tripdraw/auth/oauth/OauthClient.java
@@ -1,7 +1,7 @@
 package dev.tripdraw.auth.oauth;
 
-import dev.tripdraw.auth.domain.OauthType;
 import dev.tripdraw.auth.dto.OauthInfo;
+import dev.tripdraw.common.auth.OauthType;
 
 public interface OauthClient {
