+++ backend/src/main/java/dev/tripdraw/common/auth/Auth.java
@@ -1,4 +1,4 @@
-package dev.tripdraw.auth.presentation;
+package dev.tripdraw.common.auth;
 
 import static java.lang.annotation.ElementType.PARAMETER;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
