+++ backend/src/main/java/dev/tripdraw/post/presentation/PostController.java
@@ -5,16 +5,16 @@ import static org.springframework.http.HttpStatus.NO_CONTENT;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static org.springframework.http.MediaType.MULTIPART_FORM_DATA_VALUE;
 
-import dev.tripdraw.post.application.PostService;
+import dev.tripdraw.common.auth.Auth;
+import dev.tripdraw.common.auth.LoginUser;
 import dev.tripdraw.common.swagger.SwaggerAuthorizationRequired;
-import dev.tripdraw.auth.dto.LoginUser;
+import dev.tripdraw.post.application.PostService;
 import dev.tripdraw.post.dto.PostAndPointCreateRequest;
 import dev.tripdraw.post.dto.PostCreateResponse;
 import dev.tripdraw.post.dto.PostRequest;
 import dev.tripdraw.post.dto.PostResponse;
 import dev.tripdraw.post.dto.PostUpdateRequest;
 import dev.tripdraw.post.dto.PostsResponse;
-import dev.tripdraw.auth.presentation.Auth;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
