+++ backend/src/main/java/dev/tripdraw/trip/presentation/TripController.java
@@ -3,9 +3,10 @@ package dev.tripdraw.trip.presentation;
 import static org.springframework.http.HttpStatus.CREATED;
 import static org.springframework.http.HttpStatus.NO_CONTENT;
 
-import dev.tripdraw.trip.application.TripService;
+import dev.tripdraw.common.auth.Auth;
+import dev.tripdraw.common.auth.LoginUser;
 import dev.tripdraw.common.swagger.SwaggerAuthorizationRequired;
-import dev.tripdraw.auth.dto.LoginUser;
+import dev.tripdraw.trip.application.TripService;
 import dev.tripdraw.trip.dto.PointCreateRequest;
 import dev.tripdraw.trip.dto.PointCreateResponse;
 import dev.tripdraw.trip.dto.PointResponse;
@@ -13,7 +14,6 @@ import dev.tripdraw.trip.dto.TripCreateResponse;
 import dev.tripdraw.trip.dto.TripResponse;
 import dev.tripdraw.trip.dto.TripUpdateRequest;
 import dev.tripdraw.trip.dto.TripsSearchResponse;
-import dev.tripdraw.auth.presentation.Auth;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
