+++ backend/src/test/java/dev/tripdraw/auth/presentation/AuthControllerTest.java
@@ -1,6 +1,6 @@
 package dev.tripdraw.auth.presentation;
 
-import static dev.tripdraw.auth.domain.OauthType.KAKAO;
+import static dev.tripdraw.common.auth.OauthType.KAKAO;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import static org.mockito.Mockito.when;
 import static org.springframework.http.HttpStatus.BAD_REQUEST;
