+++ backend/src/test/java/dev/tripdraw/member/domain/MemberTest.java
@@ -1,6 +1,6 @@
 package dev.tripdraw.member.domain;
 
-import static dev.tripdraw.auth.domain.OauthType.KAKAO;
+import static dev.tripdraw.common.auth.OauthType.KAKAO;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayNameGeneration;
