+++ backend/src/test/java/dev/tripdraw/member/presentation/MemberControllerTest.java
@@ -1,6 +1,6 @@
 package dev.tripdraw.member.presentation;
 
-import static dev.tripdraw.auth.domain.OauthType.KAKAO;
+import static dev.tripdraw.common.auth.OauthType.KAKAO;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import static org.springframework.http.HttpStatus.FORBIDDEN;
 import static org.springframework.http.HttpStatus.NOT_FOUND;
