+++ backend/src/test/java/dev/tripdraw/test/TestKakaoApiClient.java
@@ -1,10 +1,10 @@
 package dev.tripdraw.test;
 
-import static dev.tripdraw.auth.domain.OauthType.KAKAO;
+import static dev.tripdraw.common.auth.OauthType.KAKAO;
 
-import dev.tripdraw.auth.domain.OauthType;
 import dev.tripdraw.auth.dto.OauthInfo;
 import dev.tripdraw.auth.oauth.OauthClient;
+import dev.tripdraw.common.auth.OauthType;
 
 public class TestKakaoApiClient implements OauthClient {
